+++ test/libsolidity/smtCheckerTests/inline_assembly/empty.sol
@@ -1,0 +1,11 @@
+pragma experimental SMTChecker;
+
+contract C
+{
+	function f() public pure {
+		assembly {
+		}
+	}
+}
+// ----
+// Warning: (76-93): Assertion checker does not support inline assembly.